Changing from oauth2 to generic
