+++ CHANGELOG.md
@@ -28,7 +28,7 @@ command line for details.
 
 - Enable username normalization (for mixed-case names on GitHub, requires JupyterHub 0.5).
   This removes `GitHubOAuthenticator.username_map` introduced in 0.3,
-  because the base Authenticator has `.username_map` as of 0.5.
+  because the oauth2 Authenticator has `.username_map` as of 0.5.
 
 ## [0.3] - 2016-04-20
